Create Menu.cpp
Here is the Menu code. There is a Menu with 4 options, the 1st one starts a game, the 2nd one shows you the rules, the 3rd one let you choose how many players are participating in the game and how many are computer players (this doesn't actually work, you can select the number of the players and the computer players but it doesn't change anything yet, it is for future reference) and the 4th is exit. I also added replayability (you can keep selecting new game untill you choose the exit option).